add falling with gravity

+++ app/src/main/jni/core/Clock.h
@@ -7,6 +7,6 @@
 using Clock = std::chrono::high_resolution_clock;
 using Time = std::chrono::time_point<Clock>;
 using ClockDuration = Clock::duration;
-using FrameDuration = std::chrono::duration<int64_t , std::ratio<1, 60>>; // `1/60` of second
+using FrameDuration = std::chrono::duration<float , std::ratio<1, 60>>; // `1/60` of second
 
 #endif //FLAPPY_CAT_ANDROIDCLOCK_H
